+++ src/Command/AbstractCommand.php
@@ -556,12 +556,10 @@ abstract class AbstractCommand implements CommandInterface
     /**
      * Executes a prepared statement.
      *
-     * @param string|null $rawSql Deprecated. Use `null` value. Will be removed in version 2.0.0.
-     *
      * @throws Exception
      * @throws Throwable
      */
-    abstract protected function internalExecute(string|null $rawSql): void;
+    abstract protected function internalExecute(): void;
 
     /**
      * Check if the value has a given flag.
@@ -589,7 +587,7 @@ abstract class AbstractCommand implements CommandInterface
         $isReadMode = $this->isReadMode($queryMode);
         $this->prepare($isReadMode);
 
-        $this->internalExecute(null);
+        $this->internalExecute();
 
         /** @psalm-var mixed $result */
         $result = $this->internalGetQueryResult($queryMode);