+++ CHANGELOG.md
@@ -19,6 +19,8 @@
 - Chg #839: Remove `TableSchemaInterface::compositeForeignKey()` method (@Tigrov)
 - Chg #840: Remove parameter `$withColumn` from `QuoterInterface::getTableNameParts()` method (@Tigrov)
 - Chg #840: Remove `Quoter::unquoteParts()` method (@Tigrov)
+- Chg #841: Remove `$rawSql` parameter from `AbstractCommand::internalExecute()` method 
+  and `AbstractPdoCommand::internalExecute()` method (@Tigrov)
 
 ## 1.3.0 March 21, 2024